+++ ModernApps/Northwind.Maui.Client/App.xaml.cs
@@ -9,6 +9,7 @@ namespace Northwind.Maui.Client
       try
       {
         InitializeComponent();
+        MainPage = new AppShell();
       }
       catch (Exception ex)
       {