+++ ModernApps/Northwind.Maui.Client/CategoriesPage.xaml.xaml.cs
@@ -1,0 +1,9 @@
+namespace MauiApp1;
+
+public partial class CategoriesPage.xaml : ContentPage
+{
+	public CategoriesPage.xaml()
+	{
+		InitializeComponent();
+	}
+}