+++ ModernApps/Northwind.Maui.Client/CustomerDetailPage.xaml.xaml.cs
@@ -1,0 +1,9 @@
+namespace MauiApp1;
+
+public partial class CustomerDetailPage.xaml : ContentPage
+{
+	public CustomerDetailPage.xaml()
+	{
+		InitializeComponent();
+	}
+}