+++ ModernApps/Northwind.Maui.Client/CustomersPage.xaml.xaml.cs
@@ -1,0 +1,9 @@
+namespace MauiApp1;
+
+public partial class CustomersPage.xaml : ContentPage
+{
+	public CustomersPage.xaml()
+	{
+		InitializeComponent();
+	}
+}