+++ ModernApps/Northwind.Maui.Client/EmployeesPage.xaml.xaml.cs
@@ -1,0 +1,9 @@
+namespace MauiApp1;
+
+public partial class EmployeesPage.xaml : ContentPage
+{
+	public EmployeesPage.xaml()
+	{
+		InitializeComponent();
+	}
+}