+++ ModernApps/Northwind.Maui.Client/ProductsPage.xaml.xaml.cs
@@ -1,0 +1,9 @@
+namespace MauiApp1;
+
+public partial class ProductsPage.xaml : ContentPage
+{
+	public ProductsPage.xaml()
+	{
+		InitializeComponent();
+	}
+}