+++ ModernApps/Northwind.Maui.Client/SettingsPage.xaml.xaml.cs
@@ -1,0 +1,9 @@
+namespace MauiApp1;
+
+public partial class SettingsPage.xaml : ContentPage
+{
+	public SettingsPage.xaml()
+	{
+		InitializeComponent();
+	}
+}